Move Room and make_map() to maps.py
